Diversos ajustes para contemplar gráficos em relatórios html

+++ src/SME.SR.Infra/Dtos/Relatorios/Comuns/GraficoBarrasVerticalEixoYDto.cs
@@ -4,7 +4,7 @@
     {
 
 
-        public GraficoBarrasVerticalEixoYDto(decimal valorTotalAltura, string titulo, decimal valorMaximoEixo, int divisoesDoEixo)
+        public GraficoBarrasVerticalEixoYDto(decimal valorTotalAltura, string titulo, int valorMaximoEixo, int divisoesDoEixo)
         {
             ValorTotalAltura = valorTotalAltura;
             ValorMaximoEixo = valorMaximoEixo;
@@ -14,7 +14,7 @@
 
         public decimal ValorTotalAltura { get; set; }
         public string Titulo { get; set; }        
-        public decimal ValorMaximoEixo { get; set; }
+        public int ValorMaximoEixo { get; set; }
         public int DivisoesDoEixo { get; set; }
 
         public decimal ValorIncrementoDivisao { get { return ValorMaximoEixo / DivisoesDoEixo; } }
